printed reports

+++ database/migrations/2020_03_18_043753_fee_structures.php
@@ -18,7 +18,7 @@ class FeeStructures extends Migration
             $table->year('year');
             $table->string('class');
             $table->integer('term');
-            $table->double('fee', 2);
+            $table->decimal('fee', 15, 2);
             $table->timestamps();
         });
     }
